Replaced Imaginary Teleprompter with QPrompt
I'm Imaginary Teleprompter's lead developer and co-author. For the past year I've worked on a new teleprompter software that doesn't have the architectural deficiencies of Imaginary Teleprompter. The new program, called QPrompt, is meant to supersede Imaginary Teleprompter as desktop software. QPrompt is currently in beta, but it is set to come out of beta in a matter of days.

A few distinctions that make QPrompt more awesome than Imaginary Teleprompter are:
 * Works with both studio and tablet teleprompters
 * Fluid motion, jitter free experience
 * Paste from other software without hassle
 * Make changes from the prompter itself, while prompting
 * Prompt to multiple screens, with independent mirroring
 * Background transparency allows you to monitor yourself or your audience as you speak
 * Estimates remaining time for you
 * Progress indicator
 * Supports writing systems of 182 languages